+++ applications/fishing-map/src/features/dataviews/dataviews.utils.ts
@@ -1,21 +1,23 @@
+import { kebabCase } from 'lodash'
 import {
   ColorCyclingType,
   Dataset,
+  Dataview,
   DataviewCategory,
   DataviewDatasetConfig,
   DataviewInstance,
   EndpointId,
 } from '@globalfishingwatch/api-types'
 import { Generators } from '@globalfishingwatch/layer-composer'
+import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client/dist'
 import {
   TEMPLATE_ENVIRONMENT_DATAVIEW_ID,
   TEMPLATE_CONTEXT_DATAVIEW_ID,
   DEFAULT_FISHING_DATAVIEW_ID,
   TEMPLATE_VESSEL_DATAVIEW_ID,
-  DEFAULT_PRESENCE_DATAVIEW_ID,
   TEMPLATE_USER_TRACK_ID,
-  DEFAULT_VIIRS_DATAVIEW_ID,
 } from 'data/workspaces'
+import { isPrivateDataset } from 'features/datasets/datasets.utils'
 
 // used in workspaces with encounter events layers
 export const ENCOUNTER_EVENTS_SOURCE_ID = 'encounter-events'
@@ -23,31 +25,32 @@ export const FISHING_LAYER_PREFIX = 'fishing-'
 export const VESSEL_LAYER_PREFIX = 'vessel-'
 export const ENVIRONMENTAL_LAYER_PREFIX = 'environment-'
 export const CONTEXT_LAYER_PREFIX = 'context-'
-export const PRESENCE_LAYER_ID = 'presence'
-export const VIIRS_LAYER_ID = 'presence'
 export const VESSEL_DATAVIEW_INSTANCE_PREFIX = 'vessel-'
 
 type VesselInstanceDatasets = {
-  trackDatasetId: string
-  infoDatasetId: string
+  trackDatasetId?: string
+  infoDatasetId?: string
   eventsDatasetsId?: string[]
 }
 export const getVesselDataviewInstance = (
   vessel: { id: string },
   { trackDatasetId, infoDatasetId, eventsDatasetsId }: VesselInstanceDatasets
 ): DataviewInstance<Generators.Type> => {
-  const datasetsConfig: DataviewDatasetConfig[] = [
-    {
-      datasetId: trackDatasetId,
-      params: [{ id: 'vesselId', value: vessel.id }],
-      endpoint: EndpointId.Tracks,
-    },
-    {
+  const datasetsConfig: DataviewDatasetConfig[] = []
+  if (infoDatasetId) {
+    datasetsConfig.push({
       datasetId: infoDatasetId,
       params: [{ id: 'vesselId', value: vessel.id }],
       endpoint: EndpointId.Vessel,
-    },
-  ]
+    })
+  }
+  if (trackDatasetId) {
+    datasetsConfig.push({
+      datasetId: trackDatasetId,
+      params: [{ id: 'vesselId', value: vessel.id }],
+      endpoint: EndpointId.Tracks,
+    })
+  }
   if (eventsDatasetsId) {
     eventsDatasetsId.forEach((eventDatasetId) => {
       datasetsConfig.push({
@@ -78,26 +81,6 @@ export const getFishingDataviewInstance = (): DataviewInstance<Generators.Type> 
       colorCyclingType: 'fill' as ColorCyclingType,
     },
     dataviewId: DEFAULT_FISHING_DATAVIEW_ID,
-  }
-}
-
-export const getPresenceDataviewInstance = (): DataviewInstance<Generators.Type> => {
-  return {
-    id: `${PRESENCE_LAYER_ID}-${Date.now()}`,
-    config: {
-      colorCyclingType: 'fill' as ColorCyclingType,
-    },
-    dataviewId: DEFAULT_PRESENCE_DATAVIEW_ID,
-  }
-}
-
-export const getViirsDataviewInstance = (): DataviewInstance<Generators.Type> => {
-  return {
-    id: `${VIIRS_LAYER_ID}-${Date.now()}`,
-    config: {
-      colorCyclingType: 'fill' as ColorCyclingType,
-    },
-    dataviewId: DEFAULT_VIIRS_DATAVIEW_ID,
   }
 }
 
@@ -160,4 +143,29 @@ export const getContextDataviewInstance = (
     ],
   }
   return contextDataviewInstance
+}
+
+export const getDataviewInstanceFromDataview = (dataview: Dataview) => {
+  return {
+    id: `${kebabCase(dataview.name)}-${Date.now()}`,
+    dataviewId: dataview.id,
+  }
+}
+
+export const getActivityDataviewInstanceFromDataview = (
+  dataview?: Dataview
+): DataviewInstance<Generators.Type> | undefined => {
+  if (!dataview) return
+  const instance = getDataviewInstanceFromDataview(dataview)
+  return {
+    ...instance,
+    config: {
+      colorCyclingType: 'fill' as ColorCyclingType,
+    },
+  }
+}
+
+export const dataviewWithPrivateDatasets = (dataview: UrlDataviewInstance) => {
+  const datasets = dataview.datasets || []
+  return datasets.some(isPrivateDataset)
 }