+++ applications/fishing-map/src/features/workspace/Workspace.tsx
@@ -19,7 +19,7 @@ import { updateLocation } from 'routes/routes.actions'
 import LocalStorageLoginLink from 'routes/LoginLink'
 import { logoutUserThunk, selectUserData } from 'features/user/user.slice'
 import { selectReadOnly, selectSearchQuery } from 'features/app/app.selectors'
-import { SUPPORT_EMAIL } from 'data/config'
+import { PRIVATE_SUFIX, SUPPORT_EMAIL } from 'data/config'
 import { WorkspaceCategories } from 'data/workspaces'
 import { selectDataviewsResourceQueries } from 'features/dataviews/dataviews.selectors'
 import ActivitySection from './activity/ActivitySection'
@@ -152,7 +152,12 @@ function Workspace() {
       {(locationCategory === WorkspaceCategories.MarineManager ||
         locationCategory === WorkspaceCategories.FishingActivity) &&
         workspace?.name &&
-        !readOnly && <h2 className={styles.title}>{workspace.name}</h2>}
+        !readOnly && (
+          <h2 className={styles.title}>
+            {workspace.id.startsWith(PRIVATE_SUFIX) && '🔒 '}
+            {workspace.name}
+          </h2>
+        )}
       <ActivitySection />
       <VesselsSection />
       <EventsSection />