+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @globalfishingwatchapp/fishing-map
 
+## 1.1.29
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/ui-components@10.1.2
+
+## 1.1.28
+
+### Patch Changes
+
+- Updated dependencies [eae1524c]
+  - @globalfishingwatch/ui-components@10.1.1
+
+## 1.1.27
+
+### Patch Changes
+
+- Updated dependencies [5fccc9f4]
+- Updated dependencies [06bbf6b9]
+  - @globalfishingwatch/api-types@5.1.0
+  - @globalfishingwatch/layer-composer@9.1.0
+  - @globalfishingwatch/ui-components@10.1.0
+
 ## 1.1.26
 
 ### Patch Changes